app/iam/opal-dynamic-policy/requirements.txt
# Combined requirements for pep-proxy + bundle-server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
httpx==0.25.1
pydantic==2.4.2
python-multipart==0.0.6
aiofiles==23.2.1
pyyaml==6.0.1
# PostgreSQL async driver (policy persistence)
asyncpg==0.29.0
# gRPC ext-authz server (agentgateway integration)
grpcio==1.59.3
grpcio-tools==1.59.3
